fix: extracted vars from component export
